create a folder models and separate interface into package and country
