+++ src/app/models/Country.ts
@@ -1,0 +1,8 @@
+import { Package } from "./Package";
+
+export interface Country {
+  countryCode: string;
+  countryName: string;
+  packages: Package[];
+  available: boolean;
+}
